+++ src/Transport/Bus.java
@@ -1,0 +1,12 @@
+package Transport;
+
+public class Bus extends Transport {
+    public Bus(String brand, String model, int year, String color, String country, int maxSpeed) {
+        super(brand, model, year, color, country, maxSpeed);
+    }
+
+    @Override
+    public String toString() {
+        return "Марка - " + getBrand() + ", Модель - " + getModel() + ", Цвет авто - " + getColor() + ", Год производства - " + getYear() + ", Страна производства - " + getCountry();
+    }
+}
